[ska] relocate saved_object_tagging (#227887)
## Summary

Part of https://github.com/elastic/kibana-team/issues/1503

This PR is mostly about moving platform tests from x-pack/test

```
x-pack/test/
     | - saved_object_tagging
```

After:
```
x-pack/platform/test/
     | - saved_object_tagging
```

(cherry picked from commit ab6be060106737a4e3c5abfc732114339867c890)
