+++ x-pack/platform/test/functional/page_objects/index.ts
@@ -26,6 +26,7 @@ import { GraphPageObject } from './graph_page';
 import { MaintenanceWindowsPageProvider } from './maintenance_windows_page';
 import { BannersPageObject } from './banners_page';
 import { NavigationalSearchPageObject } from './navigational_search';
+import { TagManagementPageObject } from './tag_management_page';
 
 // just like services, PageObjects are defined as a map of
 // names to Providers. Merge in Kibana's or pick specific ones
@@ -51,4 +52,5 @@ export const pageObjects = {
   maintenanceWindows: MaintenanceWindowsPageProvider,
   banners: BannersPageObject,
   navigationalSearch: NavigationalSearchPageObject,
+  tagManagement: TagManagementPageObject,
 };
