+++ x-pack/platform/test/saved_object_tagging/api_integration/security_and_spaces/apis/_bulk_delete.ts
