+++ x-pack/platform/test/saved_object_tagging/api_integration/security_and_spaces/apis/_get_assignable_types.ts
