+++ x-pack/platform/test/saved_object_tagging/common/fixtures/es_archiver/bulk_assign/data.json
