+++ x-pack/platform/test/saved_object_tagging/common/fixtures/es_archiver/delete_with_references/data.json
