+++ x-pack/platform/test/saved_object_tagging/common/fixtures/es_archiver/functional_base/data.json
