+++ x-pack/platform/test/saved_object_tagging/common/fixtures/es_archiver/logstash_functional/mappings.json
