+++ x-pack/platform/test/saved_object_tagging/common/fixtures/es_archiver/rbac_tags/default_space.json
