+++ x-pack/platform/test/saved_object_tagging/common/lib/authentication.ts
