+++ x-pack/platform/test/saved_object_tagging/functional/config.ts
@@ -12,7 +12,7 @@ import { services, pageObjects } from './ftr_provider_context';
 // eslint-disable-next-line import/no-default-export
 export default async function ({ readConfigFile }: FtrConfigProviderContext) {
   const kibanaFunctionalConfig = await readConfigFile(
-    require.resolve('../../functional/config.base.js')
+    require.resolve('../../functional/config.base.ts')
   );
 
   return {
