+++ x-pack/platform/test/saved_object_tagging/functional/tests/bulk_actions.ts
@@ -17,13 +17,13 @@ export default function ({ getPageObjects, getService }: FtrProviderContext) {
   describe('table bulk actions', () => {
     beforeEach(async () => {
       await kibanaServer.importExport.load(
-        'x-pack/test/saved_object_tagging/common/fixtures/es_archiver/functional_base/data.json'
+        'x-pack/platform/test/saved_object_tagging/common/fixtures/es_archiver/functional_base/data.json'
       );
       await tagManagementPage.navigateTo();
     });
     afterEach(async () => {
       await kibanaServer.importExport.unload(
-        'x-pack/test/saved_object_tagging/common/fixtures/es_archiver/functional_base/data.json'
+        'x-pack/platform/test/saved_object_tagging/common/fixtures/es_archiver/functional_base/data.json'
       );
     });
