+++ x-pack/platform/test/saved_object_tagging/functional/tests/listing.ts
@@ -17,13 +17,13 @@ export default function ({ getPageObjects, getService }: FtrProviderContext) {
   describe('table listing', () => {
     before(async () => {
       await kibanaServer.importExport.load(
-        'x-pack/test/saved_object_tagging/common/fixtures/es_archiver/functional_base/data.json'
+        'x-pack/platform/test/saved_object_tagging/common/fixtures/es_archiver/functional_base/data.json'
       );
       await tagManagementPage.navigateTo();
     });
     after(async () => {
       await kibanaServer.importExport.unload(
-        'x-pack/test/saved_object_tagging/common/fixtures/es_archiver/functional_base/data.json'
+        'x-pack/platform/test/saved_object_tagging/common/fixtures/es_archiver/functional_base/data.json'
       );
     });
